app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
aiohttp
pdfplumber
sentence-transformers
faiss-cpu
groq
tiktoken
python-dotenv
pypdf
rank-bm25
numpy
torch
